// src/app/core/models/marketing.model.ts
export interface MarketingPayload {
  PublicCible: string;
  TrancheAge: string;
  StyleDeVie: string;
  OccasionPortee: string;
  NiveauGamme: number | null;
  NoteAttractiviteVisuelle: number | null;
  PrixVenteEstime: number | null;
  PrixPsychologique: number | null;
  QuantiteEstimee: number | null;
  IndiceCompetitivite: number | null;
  MargeCible: number | null;
  USP_ArgumentUnique: string;
  ReferenceBestSeller: string;
  NomCommercial: string;
  CanalDistribution: string;
  ArgumentSecondeVie: string;
  ScoreEcoConception: string;
  Forces: string;
  Faiblesses: string;
  Opportunites: string;
  Menaces: string;
}